cleanup more from test/backend
